テストを1つに統合
